refactor: 블로그 리포지토리 삭제
