+++ src/core/blogs/blog.service.module.ts
@@ -1,11 +1,10 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { BlogService } from './blog.service';
-import { BlogRepository } from './repository/blog.repository';
 import { TaskServiceModule } from '../task/taskService.module';
 
 @Module({
     imports: [forwardRef(() => TaskServiceModule)],
-    providers: [BlogService, BlogRepository],
-    exports: [BlogService, BlogRepository],
+    providers: [BlogService],
+    exports: [BlogService],
 })
 export class BlogServiceModule {}
