+++ src/core/task/taskService.module.ts
@@ -1,9 +1,9 @@
 import { Module, forwardRef } from '@nestjs/common';
-import { TaskService } from './task.service';
 import { RedisModule } from '../../infra/redis/redis.module';
 import { RabbitMQModule } from '../../infra/rabbitmq/rabbitmq.module';
 import { BlogServiceModule } from '../blogs/blog.service.module';
 import { ScheduleModule } from '@nestjs/schedule';
+import { TaskService } from './task.service';
 
 @Module({
     imports: [
